+++ android/src/main/java/com/byteowls/capacitor/oauth2/ConfigUtils.java
@@ -1,7 +1,7 @@
 package com.byteowls.capacitor.oauth2;
 
 import com.getcapacitor.JSObject;
-import com.getcapacitor.PluginCall;
+
 import org.json.JSONObject;
 
 import java.util.HashMap;
@@ -14,18 +14,18 @@ import java.util.Random;
  */
 public abstract class ConfigUtils {
 
-    public static String getCallString(PluginCall call, String key) {
-        return getCallParam(String.class, call, key);
+    public static String getParamString(JSObject data, String key) {
+        return getParam(String.class, data, key);
     }
 
-    public static <T> T getCallParam(Class<T> clazz, PluginCall call, String key) {
-        return getCallParam(clazz, call, key, null);
+    public static <T> T getParam(Class<T> clazz, JSObject data, String key) {
+        return getParam(clazz, data, key, null);
     }
 
-    public static <T> T getCallParam(Class<T> clazz, PluginCall call, String key, T defaultValue) {
+    public static <T> T getParam(Class<T> clazz, JSObject data, String key, T defaultValue) {
         String k = getDeepestKey(key);
         try {
-            JSONObject o = getDeepestObject(call.getData(), key);
+            JSONObject o = getDeepestObject(data, key);
 
             Object value = null;
             if (clazz.isAssignableFrom(String.class)) {
@@ -48,32 +48,31 @@ public abstract class ConfigUtils {
                 return defaultValue;
             }
             return (T) value;
-        } catch (Exception ignore) {}
+        } catch (Exception ignore) {
+        }
         return defaultValue;
     }
 
-    public static Map<String, String> getCallParamMap(PluginCall call, String key) {
+    public static Map<String, String> getParamMap(JSObject data, String key) {
         Map<String, String> map = new HashMap<>();
         String k = getDeepestKey(key);
         try {
-            JSONObject o = getDeepestObject(call.getData(), key);
+            JSONObject o = getDeepestObject(data, key);
             JSONObject jsonObject = o.getJSONObject(k);
-            if (jsonObject != null) {
-                Iterator<String> keys = jsonObject.keys();
-                if (keys != null) {
-                    while (keys.hasNext()) {
-                        String mapKey = keys.next();
-                        if (mapKey != null && mapKey.trim().length() > 0) {
-                            String mapValue = jsonObject.getString(mapKey);
-                            if (mapValue != null && mapValue.trim().length() > 0) {
-                                map.put(mapKey, mapValue);
-                            }
+            Iterator<String> keys = jsonObject.keys();
+            if (keys != null) {
+                while (keys.hasNext()) {
+                    String mapKey = keys.next();
+                    if (mapKey != null && mapKey.trim().length() > 0) {
+                        String mapValue = jsonObject.getString(mapKey);
+                        if (mapValue != null) {
+                            map.put(mapKey, mapValue);
                         }
                     }
                 }
-
             }
-        } catch (Exception ignore) {}
+        } catch (Exception ignore) {
+        }
         return map;
     }
 
@@ -89,27 +88,51 @@ public abstract class ConfigUtils {
         // Split on periods
         String[] parts = key.split("\\.");
         // Search until the second to last part of the key
-        for (int i = 0; i < parts.length-1; i++) {
+        for (int i = 0; i < parts.length - 1; i++) {
             String k = parts[i];
             o = o.getJSObject(k);
         }
         return o;
     }
 
+    public static <T> T getOverwrittenAndroidParam(Class<T> clazz, JSObject data, String key) {
+        T baseParam = getParam(clazz, data, key);
+        T androidParam = getParam(clazz, data, "android." + key);
+        if (androidParam != null) {
+            baseParam = androidParam;
+        }
+        return baseParam;
+    }
+
+    public static Map<String, String> getOverwrittenAndroidParamMap(JSObject data, String key) {
+        Map<String, String> baseParam = getParamMap(data, key);
+        Map<String, String> androidParam = getParamMap(data, "android." + key);
+        Map<String, String> mergedParam = new HashMap<>(baseParam);
+        mergedParam.putAll(androidParam);
+        return mergedParam;
+    }
+
     public static String getRandomString(int len) {
-        char[] ch = { '0', '1', '2', '3', '4', '5', '6', '7', '8', '9',
+        char[] ch = {'0', '1', '2', '3', '4', '5', '6', '7', '8', '9',
             'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L',
             'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X',
             'Y', 'Z', 'a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j',
             'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v',
-            'w', 'x', 'y', 'z' };
+            'w', 'x', 'y', 'z'};
 
         char[] c = new char[len];
         Random random = new Random();
         for (int i = 0; i < len; i++) {
-            c[i]=ch[random.nextInt(ch.length)];
+            c[i] = ch[random.nextInt(ch.length)];
         }
         return new String(c);
+    }
+
+    public static String trimToNull(String value) {
+        if (value != null && value.trim().length() == 0) {
+            return null;
+        }
+        return value;
     }